fix #4 split .h in .h and .cpp (#5)
