+++ library.properties
@@ -1,5 +1,5 @@
 name=Fletcher
-version=0.1.2
+version=0.1.3
 author=Rob Tillaart <rob.tillaart@gmail.com>
 maintainer=Rob Tillaart <rob.tillaart@gmail.com>
 sentence="Arduino Library for calculating Fletcher's checksum.
